# Disable undo during bulk testing