@eevee 2020-12-20
Holy cow, this had crossed my mind but I didn't expect that much of a difference.  18% isn't exactly "barely moves"